app/requirements.txt
fastapi==0.111.0
faiss-cpu==1.12.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
langchain==0.2.0
langchain-openai==0.1.7
langchain-community==0.2.0
#faiss-cpu==1.8.0
openai==1.30.1
python-dotenv==1.0.1
boto3==1.34.100
